+++ hyperdbg/libhyperdbg/code/debugger/commands/debugging-commands/output.cpp
@@ -28,10 +28,12 @@ CommandOutputHelp()
     ShowMessages("output : creates an output instance that can be used in event "
                  "forwarding.\n\n");
 
+    ShowMessages("syntax : \toutput\n");
     ShowMessages("syntax : \toutput [create Name (string)] [file|namedpipe|tcp|module Address (string)]\n");
     ShowMessages("syntax : \toutput [open|close Name (string)]\n");
 
     ShowMessages("\n");
+    ShowMessages("\t\te.g : output\n");
     ShowMessages("\t\te.g : output create MyOutputName1 file "
                  "c:\\rev\\output.txt\n");
     ShowMessages("\t\te.g : output create MyOutputName1 file "